% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{fct_load_shovel_df}
\alias{fct_load_shovel_df}
\title{Load shovel checklist data from googlesheet and set row names}
\usage{
fct_load_shovel_df()
}
\description{
Load shovel checklist data from googlesheet and set row names
}
