% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shovelApp-package.R
\docType{package}
\name{shovelApp-package}
\alias{shovelApp}
\alias{shovelApp-package}
\title{shovelApp: Snow Shovel Routes Checklist}
\description{
A summary of whether shovel routes were completed. The data comes from Qualtrics route checklists that the snow shovelers fill out after they complete their routes each day, even if no action is required. Maria Dimacuha (mdimac01@uoguelph.ca) used to manually check the data and sends this information out at noon everyday from December to March. It’s simply a ‘pass/fail’ compliance report.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asadow/shovelApp}
  \item Report bugs at \url{https://github.com/asadow/shovelApp/issues}
}

}
\author{
\strong{Maintainer}: Adam Sadowski \email{asadowsk@uoguelph.ca}

}
\keyword{internal}
