% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{utils_grey_if_na}
\alias{utils_grey_if_na}
\title{Give a list of grey background values if missing}
\usage{
utils_grey_if_na(value)
}
\arguments{
\item{value}{Value of table cell.}
}
\description{
The list given by `utils_grey_if_na()` is to be used
in the `style` argument of `reactable::colDef()`.
}
