% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{utils_named_map}
\alias{utils_named_map}
\title{`set_names()` then `map()`}
\usage{
utils_named_map(x, .f)
}
\arguments{
\item{x}{An input.}

\item{.f}{A function for `map()`.}
}
\description{
This combination is commonly used on a vector of strings
(columns names in a `reactable`) to apply `reactable` column definitions.
}
